LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-import-data-url.any-expected.txt

Blocked access to external URL https://www1.localhost:9443/workers/modules/resources/export-on-load-script.py
CONSOLE MESSAGE: Importing a module script failed.

Harness Error (FAIL), message = Script error.

PASS Static import.
TIMEOUT Static import (cross-origin). Test timed out
NOTRUN Static import (redirect).
NOTRUN Nested static import.
NOTRUN Static import and then dynamic import.
NOTRUN Dynamic import.
NOTRUN Nested dynamic import.
NOTRUN Dynamic import and then static import.
NOTRUN eval(import()).
